refactor-ir-cake: Scalan moved to sigma.compiler package
