+++ sc/js/src/test/scala/scalan/Platform.scala
@@ -1,5 +1,7 @@
 package scalan
 
+import sigma.compiler.Scalan
+
 import scala.annotation.unused
 
 object Platform {
