+++ sc/shared/src/main/scala/sigma/compiler/Entities.scala
@@ -1,4 +1,4 @@
-package scalan
+package sigma.compiler
 
 /** A slice in the Scalan cake with base classes for various descriptors. */
 trait Entities extends TypeDescs { self: Scalan =>
