+++ sc/shared/src/main/scala/sigma/compiler/Exceptions.scala
