+++ sc/shared/src/main/scala/sigma/compiler/ModuleInfo.scala
@@ -1,4 +1,4 @@
-package scalan
+package sigma.compiler
 
 import scalan.meta.SSymName
