+++ sc/shared/src/main/scala/sigma/compiler/Modules.scala
@@ -1,4 +1,4 @@
-package scalan
+package sigma.compiler
 
 trait Modules extends Base { self: Scalan =>
