+++ sc/shared/src/main/scala/sigma/compiler/MutableLazy.scala
