+++ sc/shared/src/main/scala/sigma/compiler/meta/SSymName.scala
