+++ sc/shared/src/main/scala/sigma/compiler/primitives/Equal.scala
@@ -1,6 +1,6 @@
-package scalan.primitives
+package sigma.compiler.primitives
 
-import scalan.{Base, Scalan}
+import sigma.compiler.{Base, Scalan}
 
 import scala.annotation.unused
