+++ sc/shared/src/main/scala/special/sigma/wrappers/WrappersModule.scala
@@ -1,6 +1,6 @@
 package sigma.wrappers
 
-import scalan.Scalan
+import sigma.compiler.Scalan
 
 trait WrappersModule
   extends special.wrappers.WrappersModule { self: Scalan => }
