+++ sc/shared/src/main/scala/wrappers/special/WSpecialPredefs.scala
@@ -1,5 +1,6 @@
 package wrappers.special {
   import scalan._
+  import sigma.compiler.{Base, Scalan}
 
   trait WSpecialPredefs extends Base { self: Scalan =>
     trait WSpecialPredef extends Def[WSpecialPredef];
