+++ sc/shared/src/test/scala/scalan/TestLibrary.scala
@@ -1,5 +1,7 @@
 package scalan
 
+import sigma.compiler.{GraphIRReflection, Scalan}
+
 trait TestLibrary extends Scalan {
   import CollBuilder._
   import SigmaDslBuilder._
